Replace random id with element id to identify forms
We use the provided JS-libraries from this plugin but we initiate them manually because we're loading comments async with ajax. So in the async loaded url we create the instance of the JS Comment object, based on the element id that the user is commenting on;

```
<script>
new Comments.Instance("#cc-w-{{ elementId }}", {
	"baseUrl": "/index.php?p=actions",
	"csrfTokenName": "{{ craft.app.config.general.csrfTokenName|e('js') }}",
	"csrfToken": "{{ craft.app.request.csrfToken|e('js') }}",
	"translations" : {
		"reply" : "{{ 'comments_reply'|t }}",
		"close" : "{{ 'comments_close'|t }}",
		"edit" : "{{ 'comments_edit'|t }}",
		"save" : "{{ 'comments_save'|t }}",
		"delete-confirm" : "{{ 'comments_delete_confirm'|t }}",
	}
	
});
</script>
```

Ergo; with a random id we can't identify the form, so I think it's better to use the unique element id. Works for us! :-)