Update CommentsController.php
When loading child comments async you still need to run some JS and grab the JS variables. Currently there is no way to do this, except loading them in a template, which is not great in some (and our) case, as we're using some async calls to load child comments.